Import py-toml-0.9.4, tweaks and ok sthen@
TOML aims to be a minimal configuration file format that's easy to read
due to obvious semantics. TOML is designed to map unambiguously to a
hash table. TOML should be easy to parse into data structures in a wide
variety of languages.

This is a Python library for parsing and creating TOML.