Import annapurna-1.203.
Annapurna is a Unicode-based font family with support for the many
diverse languages that use the Devanagari script. The design is
intended to be highly readable, reasonably compact and visually
attractive to produce quality publications. These fonts are named
after the majestic Annapurna mountain range of Nepal.

ok tb@

From George Rosamond. Thanks!