Import security/argon2
$ pkg/DESCR
C implementation of Argon2, the password-hashing function that won the Password
Hashing Competition (PHC).

Argon2 is a password-hashing function that summarizes the state of the art in
the design of memory-hard functions and can be used to hash passwords for
credential storage, key derivation, or other applications.

It has a simple design aimed at the highest memory filling rate and effective
use of multiple computing units, while still providing defense against tradeoff
attacks (by exploiting the cache and memory organization of the recent
processors).

tweaks from landry@ and sthen@. ok landry@