Build libLLVM.so and link tools with it.
This seems to be the way almost all Linux distributions and BSDs ship
LLVM and is what Mesa expects.  There are concerns this may cause issues
if ports start linking it, hopefully this won't happen as everything
links libLLVM-X.Y.so where X.Y is the branch version which will have the
same ABI throughout X.Y.Z releases.  The rust port explicitly links
statically even if shared libLLVM is available.

Patch cmake files in a similiar way to a patch in NetBSD pkgsrc to
correct symbols not getting added to the shared library on OpenBSD.

Use the documented cmake var for RTTI while here.

ok brad (MAINTAINER) sthen@ bcallah@