Hack up the autoconf check for the Ada driver so that "libm.so.10.0: minor version >= 1 expected, using it anyway" doesn't cause the port build to fail.
This problem resulted in most of the ports tree being broken on non-clang
arches and some important parts of it on clang arches too.

Similar but a bit different to a direction agreed with espie.

Really: needs new Ada bootstraps, and figuring out why base libs are
interfering with the bootstrap build environment.