a final update to geolite data 20180401.
remove the maintainer convenience target for geolite databases as there
will be no more updates to these. update MESSAGE-main.

Users should switch to geolite2 and adapt programs to using libmaxminddb
if they want to continue with the free databases. There are still a few
months left on commercial updates for geolite 1.