import os-test 0.000041 OK sthen@
Prepare to run os-test on a regular basis as a OpenBSD regression
test by providing the software as a port.  As it was not designed
to be installed as a package, implement the runtime environment in
do-install.  The port specific tool os-test uses the installed files
from /usr/local to run the tests and creates the html output in the
current directory.

Comment:
test suites for POSIX operating systems

Description:
os-test is a set of test suites for POSIX operating systems designed
to make it easy to compare differences between operating systems
and to find operating system bugs.  It consists of test suites that
focus on different operating system areas.