fix parsing the ldd output to ensure we have the correct shared-library name in the resulting .gir file. this was broken for ports where the library name matches a part of the build directory, like libgweather, libical and libgepub.
the parser would incorrectly match on the 'exe' line, which contains the temporary
object file created by g-ir-scanner with the gobject-introspection namespace identifier.
for example: libgweather-3.28.1/build-amd64/tmp-introspecteuljis0a/GWeather-3.0 instead
of libgweather-3.so.3.2.

so skip the first three lines of ldd output to prevent matching on the incorrect line
for now. it's not a pretty solution but allows further discussion with upstream on
a more correct fix.

debugged with and ok aja@