import py-cryptodomex-3.6.1
PyCryptodome is a self-contained Python package of low-level cryptographic
primitives. It is an cleaned and simplified fork of PyCrypto, exposing
almost the same API. Most applications run unmodified, apart from a very
few compatibility breaks for those parts of the API that represented a
security hazard or that were too hard to maintain.

NB: currently we're packaging cryptodomex which doesn't conflict with py-crypto.
once all callers are migrated we can switch to the regular cryptodome package.

with and ok sthen@