Import barony 3.1.4
Barony is a 3D, first-person roguelike. The goal of the game is to descend to
the bottom of a dark dungeon known as the Devil's Bastion and destroy an undead
lich named Baron Herx, who terrorized the peaceful town of Hamlet in life and is
now harboring a curse against the land from beyond the grave. To aid you in your
quest are friendly humans who have been eeking out a rough life within the
dungeon for generations, as well as any friends you can bring with you in real
life: Barony is the first of its kind as a first-person roguelike in that it
fully supports cooperative multiplayer for up to four players.

Port by David Carlier (maintainer)
Ported to OpenBSD by thfr@, bcallah@ and David Carlier
Help from sthen@, danj@, bcallah@, solene@

ok sthen@, danj@, bcallah@