update to LibreNMS 1.40 - note the new addition to the pkg-readme, repeated below.
(upstream's normal setup these days is to provide only their own code
and rely on composer to fetch the dependencies. this isn't suitable for
OS packaging so I'm pregenerating a separate distfile with dependencies
instead. unfortunately the hook to copy across existing config to the
.env file is done from composer, so that doesn't work for us, so you
have to copy it yourself.)

Updating from 1.39 or earlier
=============================
As of 1.40, LibreNMS now uses the Laravel framework which requires some
additional setup. Upstream has code to copy this across from an existing
configuration initially, but it's not functional in this packaged
environment.

Generate an APP_KEY:

# cd /var/www/librenms
# doas -u _librenms /usr/local/bin/php-7.0 artisan key:generate

Copy this to the clipboard, then edit /var/www/librenms/.env and fill in the
relevant details.

It may also be necessary to raise max_connections in /etc/my.cnf;
https://community.librenms.org/t/new-requirements-for-1-40/4057
suggests doubling this value.