Import databases/lmdb
Tweaks/Input from Raf Czlonka <rczlonka AT gmail DOT com>, sthem@, jca@
ok jca@

LMDB is an ultra-fast, ultra-compact key-value data store developed by Symas
for the OpenLDAP Project.

It uses memory-mapped files, so it has the read performance of a pure in-memory
database while still offering the persistence of standard disk-based databases,
and is only limited to the size of the virtual address space, (it is not
limited to the size of physical RAM). LMDB was originally called MDB, but was
renamed to avoid confusion with other software associated with the name MDB.