Import glide 0.13.1, from maintainer Kevin Wondratsch
Glide is a tool for managing the vendor directory within a Go package.

This feature, first introduced in Go 1.5, allows each package to have
a vendor directory containing dependent packages for the project.
These vendor packages can be installed by a tool (e.g. glide),
similar to go get or they can be vendored and distributed with the package.

ok abieber@