Import bitcoin-0.16.1
Long time ago pascal@ started working on bitcoin in openbsd-wip.
I've finished this work.

A lot of feedback, tweaks and tests from Bryan Linton <bl A shoshoni D info>,
Fabian Raetz <fabian.raetz A gmail D com>, thfr@ and kirby@. Thnks a lot!

Undefined reference to boost and db4 with GCC, so only CLANG arches are
enabled yet.

Bitcoin is an experimental new digital currency that enables instant
payments to anyone, anywhere in the world. Bitcoin uses peer-to-peer
technology to operate with no central authority: managing transactions
and issuing money are carried out collectively by the network.
Bitcoin is also the name of the open source software which enables
the use of this currency.

ok kirby@