arm64 has devel/gdb now, remove workaround and register bdep/WANTLIB
gdb is picked up because it brings libbfd, which can be used by
ocamlobjinfo to analyze (at least) .cmxs files.  Not very useful on
arm64 since native code support isn't enabled, but this is consistent
with other similar architectures.

An alternative would be to drop BFD on non-native non-dynlink archs, if
it is actually used only for .cmxs files.

Late commit because I wanted to avoid conflicting with the ongoing OCaml
update, but this one will need a refreshed diff anyway.

Tested by kettenis@