Help with the dependencies from abieber@. OK gonzalo@.
Comment:
alternative curl written in go

Description:
kurly is an alternative to the widely popular curl program.

kurly is designed to operate in a similar manner to curl, with select features.
Notably, kurly is not aiming for feature parity, but common flags and
mechanisms particularly within the HTTP(S) realm are to be expected.

kurly is a standalone binary without external dependencies.

Maintainer: The OpenBSD ports mailing-list <ports@openbsd.org>

WWW: https://gitlab.com/davidjpeacock/kurly