Import games/ezquake, a modern QuakeWorld client Original submission from Tom Murphy, who takes MAINTAINER -- thanks! ok awolk@
ezQuake is a modern QuakeWorld client focused on competitive online
play. Combining the features of modern QuakeWorld clients, ezQuake makes
QuakeWorld easier to start and play. The immortal first person shooter
Quake in a brand new skin with superb graphics and extremely fast
gameplay.

While this package and the game engine are open source, one still needs
to provide the original game assets from Quake, which is not open
source. You will need to purchase the original game to use this package.