Import py-binaryornot, a simple file content checker.
Port by Edward Lopez-Acosta, who needs it for an upcoming Cookiecutter
port.  ok sthen@

pkg/DESCR:
Ultra-lightweight pure Python package to guess whether a file is binary
or text, using a heuristic similar to Perl's `pp_fttext`. This package
is cross platform and works with a variety of file types and encodings.