Import cad/qrouter, a multi-level, over-the-cell maze router for VLSI design. Original submission from Alessandro De Laurenzis, who takes MAINTAINER -- thanks! ok bentley@
Qrouter is a tool to generate metal layers and vias to physically
connect together a netlist in a VLSI fabrication technology. It
is a maze router, otherwise known as an "over-the-cell" router or
"sea-of-gates" router. That is, unlike a channel router, it begins with
a description of placed standard cells, usually packed together at
minimum spacing, and places metal routes over the standard cells.

Qrouter uses the open standard LEF and DEF formats as file input and
output. It takes the cell definitions from a LEF file, and analyzes
the geometry for each cell to determine contact points and route
obstructions. It then reads the cell placement, pin placement, and
netlist from a DEF file, performs the detailed route, and writes an
annotated DEF file as output.