import ports/devel/samurai, updated from bcallah's earlier port, ok bcallah@
samurai is a ninja-compatible build tool written in C99 with a focus on
simplicity, speed, and portability.

samurai implements the ninja build language through version 1.8.2. It
uses the same format for .ninja_log and .ninja_deps as ninja, currently
version 5 and 3 respectively.

It is largely feature-complete and supports most of the same options as
ninja.