Remove mozilla-sync 1.1 port, unlinked since 2014.
Unlikely to come back as a port since it's now a nodejs monster, but
feel free to poke in the Attic. If someday i end up really bored^Wdrunk,
maybe i'll waste a hackathon night turning it into an ansible playbook..
but last i tried, it was a failure.

Reminded by danj@