Overhaul the TeX Live PLIST generation scripts.
Although this is a "no functional change" commit, the change does two
things:

 * Removes the dependency on py-texscythe for generating the PLISTs.
   Instead, I've added a small TLPDB parser which works (unlike
   py-texsythe) entirely in-memory. This means that generating the
   PLISTs now takes seconds instead of minutes.

 * Splits the generation of the PLISTs into two scripts. The first --
   update_plist_hints.py -- emits one line for file under the fake
   install dir, advising which PLIST the file should go in. The second
   -- write_plists.py -- reads those lines and writes them to the
   correct PLIST. Eventually write_plists.py will be killed, as we plan
   to integrate update_plist_hints.py with update-plist(1) itself.

At a later date, I will port all of these scripts to Python3, since that
seems to be the way the world is going.

The idea to integrate with update-plist(1) is espie@'s, and this change
is also OK espie@.

Many thanks.