update to 69.0.3497.100 and enable jumbo build
To improve compilation times it is possible to use Jumbo builds,
in Chromium.
The idea is to merge many source files and compile them together.
Since a large portion of Chromium's code is in shared header files
that dramatically reduces the total amount of work needed.

fix an internal build dependency while here