Revert to py-zmq-16.0.3.
py-zmq-17.1.2 breaks www/jupyter-notebook. Issue can be fixed by
updating py-ipkernel to a newer version, which does not work with the
version of py-matplotlib in ports. This can be fixed by updating
py-matplotlib to a newer version, which relies on two new ports...

For now make sure that 6.4 ships with a version of py-zmq that does not
break stuff.

OK sthen@, jca@