security update to apache httpd 2.4.35, ok giovanni@
Since we no longer have Apache 1.x there's no point renaming most of the
installed files any more, only the ones that conflict with base (httpd and
htpasswd), which avoids some tiresome hand merges that are needed in the
manpages for most updates. Courtesy symlinks added for now so that the
'xxx2' variants still work.