Free Orion is an open-source, platform independent galactic conquest game in the tradition of the Master of Orion games.
The game is currently playable, but probably only fun to the more
hard-core fans of the genre. Many features have been implemented, but
many others are not. Much of the content is provisional, and unbalanced.
In other words, it'll be a while till it is finished. See the Roadmap
(https://www.freeorion.org/index.php/Roadmap) for the big picture on our
progress.

Tweaks by sthen@, solene@, bcallah@ and me
OK solene@