Sync with base:
When merging sections into the output, lld tries to adjust the alignment of
the section to be at least as large as the entry size of the section.
This causes a later check that validates the alignment to fail if the
entry size isn't a power of two.  This happens when building some of the
java support code in ports gcc.  Fix this by sticking to the original
alignment if the entry size isn't a power of two.

from Brad (maintainer)