Import alkalami-1.100.
Alkalami is a font designed for Arabic-based writing systems in the
Kano region of Nigeria and Niger.

Alkalami (pronounced al-KA-la-mi) is the local word for the Arabic
"qalam", a type of sharpened stick used for writing on wooden boards
in the Kano region of Nigeria and in Niger, and what gives the style
its distinct appearance. The baseline stroke is very thick and
solid. The ascenders and other vertical strokes including the teeth
are very narrow when compared to the baseline. A generous line
height is necessary to allow for deep swashes and descenders, and
the overall look of the page is a very black, solid rectangle.
Diacritics are much smaller in scale, with very little distance
from the main letters.

ok sthen@

From George Rosamond (MAINTAINER); thanks!