Import blinken-18.08.2; ok bcallah@
Comment:
memory enhancement game

Description:
Blinken is based on an electronic game released in 1978, which challenges
players to remember sequences of increasing length. On the face of the device,
there are 4 different color buttons, each one with their own distinctive sound.
These buttons light up randomly, creating the sequence that the player must
then recall. If the player is successful in remembering the sequence of lights
in the correct order, he advances to the next stage, where an identical
sequence with one extra step is presented. If the player makes a mistake, the
game is lost, and the player must start again from the beginning. The goal is
to get a high score - each step in the sequence is worth one point, so correct
entry of a sequence of 8 lights is worth 8 points on the high score table.

Maintainer: KDE porting team <openbsd-kde@googlegroups.com>

WWW: https://www.kde.org/applications/education/blinken/