Import kblocks-18.08.2; tweaks and ok bcallah@
Comment:
falling blocks game

Description:
KBlocks is the classic falling blocks game. The idea is stack the
falling blocks in a way that lines are completely filled. When a
line is completed it is removed, and more space is available in the
play area. When there is not enough space for blocks to fall, the
game is over.

Maintainer: KDE porting team <openbsd-kde@googlegroups.com>

WWW: https://www.kde.org/applications/games/kblocks