Import bomber-18.08.2; tweaks and ok bcallah@
Comment:
arcade bombing game

Description:
Bomber is a single player arcade game. The player is invading various cities in
a plane that is decreasing in height.

The goal of the game is to destroy all the buildings and advance to the next
level. Each level gets a bit harder by increasing the speed of the plane and
the height of the buildings.

Maintainer: KDE porting team <openbsd-kde@googlegroups.com>

WWW: https://www.kde.org/applications/education/bomber/