Import math/pspp, a program for statistical analysis of sampled data. ok bentley@
GNU PSPP is a program for statistical analysis of sampled data. It is a
free as in freedom replacement for the proprietary program SPSS, and
appears very similar to it with a few exceptions.

The most important of these exceptions are, that there are no "time
bombs"; your copy of PSPP will not "expire" or deliberately stop working
in the future. Neither are there any artificial limits on the number of
cases or variables which you can use. There are no additional packages
to purchase in order to get "advanced" functions; all functionality that
PSPP currently supports is in the core package.

PSPP is a stable and reliable application. It can perform descriptive
statistics, T-tests, anova, linear and logistic regression, measures of
association, cluster analysis, reliability and factor analysis,
non-parametric tests and more. Its backend is designed to perform its
analyses as fast as possible, regardless of the size of the input data.
You can use PSPP with its graphical interface or the more traditional
syntax commands.