update to mupdf-1.14.0, some notable changes:
- mupdf-gl now requires building with javascript support, so is no longer
available in the default flavour (which has js disabled). If you use mupdf-gl
you will need to switch to the js flavour (pkg_delete mupdf; pkg_add mupdf--js).

- drop the patch to convert to shared libraries, the build process of
mupdf is a headache and changes rather often, so this patch is getting in
the way too much. unfortunately this bloats the packages somewhat...