Import kanagram-18.08.2
Comment:
letter order game

Description:
Kanagram is a game that is based on the word/letter puzzles that the
author played as a child. A word is picked at random and displayed with
its letters in a messed order, with difficulty dependent on the chosen
level. You have an unlimited number of attempts, and scores are kept.

Maintainer: KDE porting team <openbsd-kde@googlegroups.com>

WWW: https://edu.kde.org/kanagram/

OK landry@