Remove insecure nxssh and obsolete opennx
nxssh is based on an old copy of OpenSSH 4.7.  Upstream stopped
supporting it and it hasn't been updated since import, so it's likely
affected by all the issues listed on
https://www.openssh.com/security.html since 2008.

opennx is a graphical client that uses nxssh under the hood.  It is dead
upstream, hasn't been updated since 2009 and won't build with wxWidgets3.

ok kn@