Import net/go-ipfs.
```
IPFS is a global, versioned, peer-to-peer filesystem. It combines good ideas
from Git, BitTorrent, Kademlia, SFS, and the Web. It is like a single
bittorrent swarm, exchanging git objects. IPFS provides an interface as
simple as the HTTP web, but with permanence built in.
```

Originally from Fabian Raetz, with input from abieber@.

OK abieber@, thanks.