Import telegraf 1.9.0.
Telegraf is an agent for collecting, processing, aggregating, and
writing metrics. Plugin-driven, it has the concept of 4 distinct plugin
types:
* Input Plugins collect metrics from the system, services, or 3rd party
  APIs
* Processor Plugins transform, decorate, and/or filter metrics
* Aggregator Plugins create aggregate metrics (e.g. mean, min, max,
  quantiles, etc.)
* Output Plugins write metrics to various destinations

note that input.system is currently broken on OpenBSD (probably because
of gopsutil or go internals), tracked in
https://github.com/influxdata/telegraf/issues/5157. Other plugins seem
to work fine.

As it was done for influxdb (telegraf comes from the same source),
bundle the go deps and selfhost the distfile.

ok danj@